Initial attempt to support "postprocessors".
Creates a ProtocolPostProcessor, distinct from Protocol, which has
a 'run' method that takes an existing results/resultsdir object.  Also
refactors VolumetricBenchmarkGrid into a protocol and postprocessor as
an example.

The ProtocolPostProcessor class is essentially a copy of the Protocol
class, so it seems redundant to have a separate class for something
that is so similar to a Protocol (the only difference is that the run
method takes an results instead of a data object).  Perhaps it would
be better to just add a "postprocess" method to the Protocol class?