Adds a _loaded_from member to TreeNode for default write() calls.
TreeNode and ProtocolResults objects now have a "_loaded_from"
member that holds the path an object is loaded from, and gets
initialized by "from_dir".  The corresponding write() functions
are updated so that their 'dirname' arguments are optional, and
the stored "_loaded_from" path is used if it exists (otherwise an
error is raised).