Updates tutorial notebooks to use protocol objects.
This isn't a comprehensive update of the tutorials, but it updates
and adds to he main ones, and moves some of the replaced tutorials (which utilize
the old "function-based API") to filenames having a "-functionbased" suffix.
The thought is that we can keep these around as demonstrating a lower-level
type of accessing the Protocols.