Alters default writing directory for protocol objects.
Now only ExperimentDesign objects hold a ._loaded_from attribute,
which holds the directory (if any) the tree node was loaded from.
This allows .write() calls of data and results objects to use the
directory of their contained experiment design by default.
This is illustrated in the GST-Protocols.ipynb notebook.

This commit also fixes the inheritence order of StructuredGSTDesign
so that super() can be used correctly in all its base classes
(see the MRO for StructuredGSTDesign - now CircuitStructureDesign
has CircuitListsDesign directly below it instead of
GateSetTomographyDesign -- so super() within CircuitStructureDesign
works as expected).