Adds a bunch of minor bug fixes to get smq-modelpacks working.
This commit gets GST reports working when using the new "smq"
modelpacks.  These had issues initially because the gate names
are tuples instead of strings, so some adjustments to the report
generation needed to be made.  This commit also updates all the
_sslbls attributes in the model packs from being lists -> tuples,
since the target-model caching depends on this being hashable.