Fixes a number of small bugs; adds 'target_model' arg to ModelTest.
This commit fixes a number of bugs identified during tutorial-writing.
It adds a 'target_model' argument to ModelTest.__init__, since it
seems useful to be able to supply this at "run-time" in cases when
the experiment design doesn't have a target model (e.g. when doing
a model test on a non-GST experiment design).