Adds MarginalizedPOVM and rough auto-POVM marginalization.
A WIP, this updates the Implicit layer lizard (the explicit one
should be updated similarly) to perform all the sequence simplification
within it's __init__ rather than in the Model's _layer_lizard method.
This is preferable as it gives the layer lizard more freedome to
create operations, as in this case it now can create MarginalizedPOVM
operations when an unknown POVM label is encountered that has a base
name equal to an existing POVM.  Such marginalized-POVM labels are
given as the defaults (now) when a circuit's line-labels are less
than all the state-space (qubit) labels.  The marginalized POVMs
are currently only static objects, and so using them with GST or
in any parameterized context will break.