Upgrades do_* driver functions to use protocol objects.
In longsequence.py, the do_* driver functions (except for do_linear_gst,
which isn't implemented in Protocol objects yet) have been rewritten
to just use protocol objects.  The old versions of the functions have
been temporarily preserved with a legacy_ prefix, in case there are
bugs in the protocol objects we haven't found yet or for backward
compatibility scenarios we haven't though of.  The updated driver
functions seem to be working correctly.

This commit also updates some of the tutorials which (still) use
these driver functions to use the new smq model packs, and fixes
several bugs related to the gate labels not being simple string
(mostly due to aliases being used to translate strings and mess
up the line_labels to something like (0,'*') ).