Alters line_labels="auto" behavior in Circuit.__init__
Circuit now assigns the "old-style" ('*',) line labels to *empty*
circuits when editable=False.  Previously the empty circuit would
be automatically given no line labels (an empty tulple) so that
the ('*',) label wouldn't interfere with constructing circuits
starting with an empty one.  But in these cases one will probably
set editable=True, and in other cases, like when converting a number
of tuples-of-labels into Circuits, we want the "auto" line-labeling
to work the same for () and (label1, label2, ...).  This update
should fix this issue without breaking anything.