Makes a number of minor updates to get unit tests running.
Two tests of the "truncated germ powers" truncation method have
been skipped since StandardGSTDesign doesn't support this and there
doesn't seem to be a good reason to do so.  This commit also updates
the expected outputs of a sequential-fiducial-pair-finding test to
what appears to be an alternate, but also valid, fiducial pair set.