Finishes the WIP for updating layer lizards.
Updates the ExplicitLayerLizard to match the pattern of the
(recently updated) ImplicitLayerLizard, which internalizes the
"simplification" process whereby POVMs and instruments are converted
into effects and operations, respectively.  This means that a Model's
_layer_lizard() method no longer performs this step, and the layer
lizard has more at it's disposal when creating prep/effect/operations
on demand.  This is used by both of the implicit-model layer lizards
to create marginalized effect vectors from marginalized POVMs
(this commit only updates the one used for cloud-noise models).