Updates list_to_slice function to better facilitate custom model parameterizations.
The list_to_slice fuction's require_contiguous argument previously
raised an assertion error when the given list wasn't congiguous.  Now,
more sensibly, it only raises this error when its array_ok argument
is set to False, otherwise it can just return an array that isn't
contiguous.  This seems to be the desired behavior, as this function
is used to cast index-arrays to slices when they can be (and some
routines can only deal with contiguous slices), but when index-arrays
are ok too they don't need to be contiguous.