Updates wildcard error processing and StandardGSTDesign.
Updates the nSigma calculation used to determine whether "bad fit"
post-processing is activated.  Now this uses the number of non-gauge
model parameters instead of the total number of model parameters, as
this is what is used in the reports and the discrepancy can be
confusing.  To ensure this is robust when used with models that can't
compute non-gauge parameters, a hasattr() call checks whether the
num_nongauge_params method exists.

This commit also updates the GatesVsTarget table to properly include
"unmodeled error" for instruments.

Separately (and unrelated), this commit expands the StandardGSTDesign
so that includeLGST and nest are arguments to the constructor, giving
these experiment designs more flexibility.