Adds str(.) around `filename` in `open(filename)` for python 3.5
The issues we've been running into regarding pathlib and os.path
is that these don't play nicely together until python 3.6, and
TravisCI is testing with python3.5.  This is also true of the
open(.) function, and this commit adds explicit str(.) casts
around the filename of some open(.) calls in merge_helpers.py
and workspace.py.  When python3.5 is long gone, search for
"open(str(" to get rid of these band-aids.