Updates ExplicitModel.num_gauge_params to punt for term evotypes.
Because we updated misfit / nSigma functions to, by default, try
to call model.num_nongauge_params to obtain the parameter count
to use, this function is now called for all types of explicit
models.  This includes models with term-based evotypes, where the
implementation doesn't actually work.  For now, we just punt and
silently return 0 as the number of gauge parameters for a non
densitymx or statevec evotype explicit model.