Adds handling for LocalNoiseModel in NQubitErrgenTable.
The n-qubit version of the error generator table only worked for
cloudnoise models (it's not fully completed yet).  This commit
adds support for LocalNoiseModel objects, by printing the error
generators of their primitive layers.  In the future, more intelligent
code could be added to show, for instance, just the gate error
generators when the gates of the local noise model are not "independent".