Removes __init__.py from test/test_packages/mpi/.
The using mpinoseutils breaks when there's an __init__.py file
within test/test_packages/mpi/ because mpinoseutils uses __name__
to get the module name of test/test_packages/mpi/mpinoseutils.py.
If this module name is "test.test_packages.mpi.mpinoseutils" (this
happens when an __init__.py exists) then it causes an error because
the root "test" package directory is not on the python system path.

When there is not __init__.py, then the module names are just
"mpinoseutils" or "testmpiMain" which exist in the current working
directory (automatically in the python path) and so all works fine.

==> Thus, there should be no __init__.py in this directory, and
testmpiMain.py should use an absolute import, i.e.
"from mpinoseutils import *", not "from .mpinoseutils import *".