Added warning message to /packages/pygsti.
As of 8aaa12b, the project source was moved to the repo root. @enielse
brought up a good point that users who have installed pyGSTi as an egg
may be confused as to why their install is suddenly broken. By leaving
a small script at /packages/pygsti we can warn affected users to
reinstall pyGSTi.