Adds 'withTimes' argument to load_dataset(...)
This argument allows the user to force load_dataset to only or
never read in data using the new block-of-lines format that includes
timestamps.  The default value "auto" allows lines to be loaded in
using the old and new formats.  This extra control is particularly
useful when reading in template files that contain no count-columns,
since these lines would be interpreted as needing additional timestamp
lines if withTimes="auto" -- setting withTimes=False allows such a
file to be read in correctly.