Fixes a minor bug in Basis.cast(...)
When 'dim' is specified as an argument to cast(...) along with a Basis
object, it the dim should be allowe to be equal to the dim of the
Basis OR ot the element-size (.elsize) of the Basis.  For instance,
you should be able to cast a 2x2 + 1x1 direct-sum basis as a dim=9
basis without getting an assertion failure, since the element size
of the Basis (9) is the same as that of the basis that would have
been created.