Consolidates LGST code.
The GST protocol now calls an LGST protocol, and LGST can be used
even when an initial model is specified to the GST protocol (by
setting "starting point" = "LGST").  The "starting point" advanced
option has been updated slightly to allow a "LGST-if-possible"
value, which attempts to run LGST on the initial model (if given)
or the target model (if an initial model is not given) but just
proceeds if this fails.  By default, GST will not try to run LGST
(ever!), as the default "starting point" value is "auto".  StandardGST
*will* try to run LGST, as it uses "LGST-if-possible" as its default
option.