Updates unit test.
In a recent commit, we removed the ability to specify a Model as
the 'starting point' advanced option to GST.  This feature is now
accessible through the 'initial_model' argument of the GST protocol
object, and is no longer available as an advanced option to
do_long_sequence_gst.  Docstrings have been updated.