Fixes typo (missing kwarg name) causing bad summary histogram plots.
Since our recent reports refactor, the summary-histogram plot within
the summary report setion has been missing the "linlg_pcntle=" keyword
name in its call to ColorBoxPlot.  This caused the linlg_pcntle argument
within ColorBoxPlot to be incorrect, resulting in a historgram plot
with the wrong grayscale-to-red transition point.  This commit fixes
this issue.