Casts the axis tick labels of matrix color box plots as strings.
This allows the user to specify non-strings as the x- or y-tick
labels by casting them to strings within the matrix_color_boxplot
function.  Since data-set labels are used as the y-tick-labels of
summary goodness-of-fit plots in the standard reports, this update
is needed to allow users to specify non-string keys of result-dictionaries,
e.g. {1: results_from_dataset1, 2: results_from_dataset2}.