Removes "PDF output is not available" warnings at Report creation.
The construct_* report construction functions previously issued
warnings when they created a Report object that could not generate
PDFs because there were either multiple datasets or gauge opts.

These warnings have been removed, and the ValueError raised when
one tries to create a PDF report but cannot has been made more
descriptive.  The reasoning for this is that:
- most of the time non-PDF reports will be generated, and so these
  warnings are largely inapplicable and a nuisance to the user.
- we don't want to dissuade users from taking advantage of HTML-report
  features (e.g. the ability to present multiple dataset and gauge opts).
  Issuing these warnings may make users think they *should* always
  be creating PDF-able reports, which isn't the case.