Minor report improvements: reverts to Scatter (non-GL) and improves  dof-per-element calc for chi2 curve in histogram plot.
The Scattergl plotly plot seemed to behave badly at least in some
browsers, so we revert back to the non-WebGL version here.  Hopefully
it's not too slow.

The degrees-of-freedom per circuit used to draw a chi2 curve on the
logL histogram plot is now calculated a bit more robustly, using the
average (numOutcomes-1) over all the circuits, rather than just taking
the dataset's total nOutcomes-1 (this doesn't work when there are
circuits with different outcome sets, e.g with instruments).