Fixes bugs in Embedded and Dense Op MPI derivatives.
Fixes bugs that would result in a numpy assignment mismatch in
certain cases where MPI GST is used.  In particular, this commit
fixes the deriv_wrt_params of EmbeddedOp, which is used by a
MatrixForwardSimulator.  (The issue was that wrtFilter was applied
twice.)  This bug would only present itself when EmbeddedOps were
used with MPI and with the 'matrix' forward simulation type.

A similar issue would occur when DenseOpInterface-derived object
relied on the base-class deriv_wrt_params implementation, which
would break if used with MPI.  This bug would only present itself
in custom user-made objects, as all the derived pyGSTi objects
implement their own (faster) deriv_wrt_params methods.