Fixes a defficiency in pickling numpy arrays.
As of Numpy v1.18.1 (and maybe earlier), when arrays are pickled
and un-pickled they don't always preserve their flags.  In particular,
the "OWNDATA" flag is un-picked as False (!!) once the array size
is > 125 (seem very arbitrary).  Also, the WRITEABLE flag is un-pickled
as True even when the pickled array was read-only.  As a stopgap measure
to remedy this until numpy/pickle gets fixed, this commit adds "reducefix"
arguments to all of the dense representation objects' __init__ (or __cinit__)
functions.

Under normal circumstances "reducefix" is set or left equal to 0,
and the representation objects trust that they're being given data
that has been properly prepared.  A reducefix value of 1 or 2 indicates
that __init__ is being called to un-serialize some data, as only the
__reduce__ function should supply nonzero reducefix values.  In this
case, the __init__ function constructs a numpy array that owns its own
data and has its writeable flag set appropriately (copying the data
as needed).