Makes dense operators/spamvecs more robust to serialization.
Systematically upates operator and spamvec classes that have
dense underlying representations to ensure that instances do not
rely on multiple references (pointers) to their representation's
data.  This is to be avoided because, while the relevant __init__
methods can ensure everything gets setup correctly, building an
instance by de-serializing (e.g. un-pickling) cannot be relied
upon to keep reference "in sync".  This is accomplished by
replacing "duplicate references" with @property methods.  For
example, in a DenseOperator, instead of having a .base attribute
that __init__ sets equal to ._rep.base (and unpickling may *not*
set these to the same memory), 'base' is not a @property that
returns self._rep.base.  This update makes the operator classes
cleaner and more robust, but may also reduce performance in areas
where the code assumes that calls to operator.base is fast (maybe
in MatrixForwardSimulator?) -- in these cases it should be easy to
cache the dense operators as needed.

This commit in particular fixes an issue with the .base of a
TPSPAMVec getting out of sync with its representation's data, thereby
fixing some system tests.