Restores support for unpickling old gate sets (models).
Because recent updates changed .base and .base1D attributes into
properties, unpickling old versions of these objects was broken.
This commit updates io/legacy.py to inject hacks into the pickling
machinery when needing to load old operator objects.

This commit also updates a system test that checks a custom operation
class to actually be correct.  The test was setting the self.base
in the custom op class directly, which is now flagged as an error
(conveniently!) because 'base' is a property that cannot be set.