Attempts to improve GaugeRobustTable (from investigation of negative off-diags).
The gauge robust error metrics table seems to often have negative
off-diagonals for fidelity-like metrics, and this commit is a result
of taking a first look at why this might be.  It seems that part of
the story (but not all of it) was due to the numerical gauge optimizations
that tweak an initial "best case gauge" for a given gate.  These
gauge opts use the frobenius distance as a metric, and optimizing
this seems to result in (small) gauge tranforms that do bad things
to the fidelity.  This commit updates the table to use the fidelity
or trace distance objective functions in the gauge optimization,
depending on the quantity being reported by the table.  These
"non-standard" gauge optimizations are guarded by a try block,
which will fall back to the usual frobenius metric if anything goes
wrong, as we haven't used non-frobenius gauge optimization much at all.

This commit also adds a "metric" argument to ExplicitModel.strdiff,
so that one can easily print the distance between two explicit
models using something other than the frobenius distance.