Updates io/msgpack.py to work with current version (1.0) of msgpack.
The latest version of msgpack (v1.0) uses standard python strings,
obviating the need to convert to/from binary strings when serializing
to the msgpack format.  This is an easy update, since we use the same
non-binary-string serialization for the json format.  This commit
checks the version of msgpack and only uses the to/from binary string
conversion when necessary.