Restores setup.py to pre-debugging state after TravisCI seems fixed.
This commit gets rid of all the extra junk we've been adding to
setup.py and .travis.yml in efforts to get the unit and system
tests on TravisCI to use a regular (non editable, -e) pip install.
Problems occur because, it seems that the local git repo's path
gets added to the python system path somewhere within TravisCI's
build, causing it to import the local (not-installed) version of
pygsti which doesn't have any extensions built.  The solution is
to, after installing pygsti (via "pip install ."), rename the pygsti
directory so it isn't found when trying to "import pygsti".  This
is a pretty hacky way to get around this, but I still don't understand
how the repo path is getting into the python path so I'm not sure
how to fix this issue more elegantly.