Upgrades custom-LM routine, adding SVD and adaptive damping.
This commit adds a number of features to the LM routine in
customlm.py, allowing the algorithm to perform dampinging in the
SVD-defined basis rather than just adjusting diagaonal matrix
elements, and adding an 'adaptive' damping mode whereby the
way in which the normal equations are damped is allowed to vary
to achieve better performance.  These features are still largely
experimental, but they've proven useful enough, at least in our 4Q
test cases, to warrant implementing them as legit modes of the
optimization function.

This commit also updates core.py to be compatible with the new
optimization arguments, and also adds a bunch of commented-out
"***DEBUG***" blocks that we've been using to test different
multi-Q GST optimization strategies (still a WIP).  This commit
multiplies the user-supplied minProbClip value by 1000 to turn it
into a min-p/f-ratio value, so 1e-4 => 0.1 which is what we've been
using in our testing.