Fixes a bug in unitary gauge optimzation (used for CPTP models).
This bug fixes an bug that leads to a unitary gauge optimization
modifying the model in a non-gauge way (yikes!).  Here's what
was fixed:
- LindbladOp.spam_transform deferred to LinbladErrorGen.spam_transform,
  which is incorrect because the operation being performed is *not* a
  similarity transformation but rather a left or right multiplication.
  Now this function correctly builds a temporary LindbladOp for the
  transformed superoperator and extracts its error generator parameters.

- the LindbladPOVM's transform function would call .transform on *each*
  of its LindbladSPAMVec effect vectors, which lead to the POVM's single
  error_map being updated *multiple* times.  This bug may actually not
  have been an issue until we fixed the first bullet point above, but
  I'm not sure of this.  In any case, LindbladPOVM.transform now updates
  its .error_map *once* and just tells the effect vectors to update their
  representations (which contain pointers to the error_map).

This commit leaves a bunch of debug and alternative statements
commented out, many of which we should clean up (remove) in the next commit.

Also, some of the fixes may not work for sparse matrices - supporting the
sparse case may require some more work later on.