Fixes known issue in TP-SPAM gauge optimization on CPTP models.
The TP-SPAM gauge optimization is a 1-parameter optimization
that scales the state prep and measurement oppositely (and also
affects the non-unital parts of gates).  On CPTP models, such
transforms can scale the state prep or effect vectors too far,
making the corresponding SP or M error_map non-CP, and we would
just hope this didn't happen too much.  This commit adds a robust
fix whereby an invalid gauge transformation (one which makes an
error map non-CP and in need of truncation) is heralded and diallowed
in the optimization.

This is done by adding `oob_check_interval` arguments to gauge
optimization routines, which utilize an already-present ability of
the custom_leastsq routine.  Then, in TP-SPAM gauge optimizations
no error-map truncation is allowed (`truncate=False` in transform(...)
and spam_transform(...) methods of LindbladOp) and the potential
ValueErrors indicate an out-of-bounds region to the optimizer.