Removes the '>=1.15.0' from 'numpy>=1.15.0' in setup.py.
The TravisCI builds for python 3.5 and 3.7 are failing because
cvxpy installs but is non-importable.  This seems to be an issue
related to incompatible numpy and cvxpy versions, perhaps because
the numpy version is too old?  Maybe having the >=1.15.0 in setup.py
causes pip to *not* upgrade numpy to the latest version when there's
a stock version (like 1.15.4) available?  In any case, we probably
don't really need the >=1.15.0 restriction anyway.