Adds enhancements to ChiAlphaFunction.
Adds ability to evaluate derivative at a different x-point than
the stitch point is at, resuting in a kink at this location, but
allowing the function to not increase as drastically.  This makes
it more like our traditional chi2 function and better able to cope
with the transitions between L-value GST iterations.

This also adds an alternative to the harsh way we've been dealing
with f=0 terms in the logl and chi-alpha functions, making it again
more similar to the traditional chi2 function.  By setting radius=None
a ChiAlphaFunction will use a milder regularization of f=0 terms, using
a quadratic with curvature equal to the worst (highest) curvature used
in when stitching in regularizing parts of the f>0 terms, rather than
the cubic function dictated by a radius > 0 (that is still always used
by the logl objective function).

Finally, this commit adds a number of debug statements that we'll want
to do away with in the future but are helpful for debugging.