Removes high-x regularization in ChiAlphaFunction.
It was thought that regularizing the chi-alpha (or chi2) objective
function for x=p/f > some_threshold would be helpful in optimizations.
This wasn't found to be the case, and since the final logl function
doesn't have any such high-x features, it seems best to just abandon
this and only regularize for x < some_threshold.  This commit removes
the high-x thresholding and patching in ChiAlphaFunction.