Updates LogLFunction minProbClip naming and adds "relaxed" f=0 term mode.
Starts to switch naming over from "minProbClip" to "pfratio_stitchpt" as
this is more informative, marking with NotImplmentedError parts that still
need updating (time dependent logl and termgap mode).

Adds ability to set radius=None and use a "relaxed" mode whereby the f=0
term regularization isn't as harsh (quadratic vs cubic) and is based on
the worst (highest curvature) f>0 term regularization factors.  This commit
also adds the ability to set a different x-point at which the derivatives in
the taylor expansion regularization of a LogLFunction are take at (called 'x1'),
as this seemed a useful in the chi2 or chi-alpha case.