Fixes a loss-of-precision issue that was causing large logl jacobians.
Because of a sub-optimal order of operations, when x=p/f was very
close to 1.0 the logl objective function (v ~= sqrt(-log(x) - 1 + x)
would return 0.0 when it should have returned a small number (~1e-8
or so). Within the logl jacobian, this resulted in the "dprobs factors"
incorrectly being near 1e100, because the v=max(v,1e-100) guard left
v=1e-100 and the jacobian = (0.5/v)*(something small ~ 1e-8), as the
latter factor in the jacobian did not suffer the loss-of-precision issue.
In the end, this resulted in a falsely enormous jacobian when the
probabilities and frequences got too close to one another (x near 1.0),
which is of course where we want the optimization to get to, so this
was bad.

By simply reordering the +/- terms in v to sqrt(x - 1 - log(x)) this
problem should be fixed.  This commit takes a step further and adds
an explicit use of a Taylor expansion when |x-1| < a_threshold to
make sure things stay controlled, though this is probably not strictly
necessary.

This commit also adds a few more debug lines that were used to ferret
out the aforementioned issue.