Adds 'regtype' to LogLFunction so it can implement 2 types of regularizations.
Leading up to merging the current develop branch into this feature
branch, this commit revives some of the "old" (current-on-develop)
minProbClip-based regularization to the LogLFunction, adding a
'regtype' attribute that can either be 'minp' or 'pfratio'.  Now
the recent updates to the logl regularization/stitching are one of
the two 'regtype' modes available to LogLFunction, allowing us to
keep the older functionality and add 'pfratio' functionality as-needed
over time.  For example, the time-dependent logl function and the
'termgap' calculation functions currently only work in 'minp' mode.