Adds more experimental tests within do_iterative_mlgst.
This commit futher modifies the do_iterative_mlgst function to
test out an iterative MLGST method that uses a "lite" logl function
(using the pfratio regularization and "relaxed" f=0 term mode) instead
of a chi2 function to perform the initial optimization stages.  It
then switches to a stricter logl function, still using the pfratio
regularization but with the harsh f=0 term mode.  The do_iterative_mlgst
function is still a mess of commented out alternatives as we continue
to experiment on 4Q data.