Adds fix to boundary case of zero gauge parameters.
An ExplicitModel computes the number of gauge degrees of freedom
by taking the matrix_rank of a matrix of gauge directions.  In
certain cases there might not be any such directions, and matrix_rank
is called on a zero-length array, resulting in an error.  This case
should just return zero gauge parameters, which it does now.