Removes data-path-filtering in VolumetricBenchmarkingGrid.
Previously, the VolumetricBenchmarkingGrid.run method would filter
the ProtocolData object it was given, trimming off some of the
"paths" to unused experiment designs and data.  This is bad, however,
since the produced ProtocolResults object then contains a trimmed
ProtocolData object that, if written to the same directory as the
original un-trimmed data/edesign will ovewrite the full data with
a trimmed version.  In particular the edesign/subdirs.json file
will list fewer subdirectories than are actually present (only the
paths used by the VBGrid object).  The solution is to *not* modify
the given data & design objects within a protocol's run() method,
which is the change this commit affects.