Improves as_nameddict() and as_dataframe() for ProtocolResults.
Updates Protocol to include a setup_nameddict(...) method that
creates a nested NamedDict of attributes/options within the
Protocol (e.g. the statistical method being used).  ProtocolResults
uses this method to initialize the NamedDict returned from its
as_nameddict() method, which is in turn used to create pandas DataFrames.

In the end, this causes the DataFrames returned from as_dataframe()
to contain a lot more information about the protocol that lead to
the given results, which is beneficial for data analysis.