Adds 'quick_load' argument to protocol object loading.
Now, by setting quick_load=True to any of the various protocol-object
loading functions (e.g. load_results_from_dir), objects can be loaded
with some parts/attributes missing - those attributes that would have
taken longer to load.  This can be helpful when loading is slow and,
for instance, only the results within a ProtocolResults object are
really needed, though it's important that the user not re-write the
partially loaded object back to disk, as this could destroy existing
data.