Adds support for multi-pass data using protocol objects.
In particular, this adds support for a ProtocolData object having
a dictionary of DataSets as its .dataset instead of a MultiDataSet
(we still need to add more functionality to MultiDataSet before it
can be used).  This commit updates the MultiPassProtocol and
MultiPassResults classes, and adds a `run_passes_separately` argument
to DefaultProtocolRunner.