Substantial refactor of objective-function "tools" functions.
This commit consolidates much of the duplicated code between
objectivefns.py and chi2fns.py and likelihoodfns.py by:
- moving hessian computations (and approximate hessian computations)
  into the objective function classes (these should be tested!),
  along with other small additions.
- rewriting the tool functions to simply create and call
  ObjectiveFunction objects.
- renaming/creating tool functions, e.g. chi2_hessian is a
  new function whose work was previously done by passing a
  returnHessian=True argument to the chi2 function.

The updated tools function still need to be tested, and we should
probably update their argument ordering to have some consistency
with the objectivefn classes too.  Then we'll need to update the
functions which call these, to fix naming and argument ordering.

Further consolidation could be performed by organizing the
objective fn classes a little more, and separating out the model-requiring
part of the computations so as to also expose a "bare" objective
function method that can replace or implement the tool functions such
as chi2_fn or two_delta_logl_fn that operate on probabilities and
counts directly.