Adds an additional check to operation.convert() for CPTP gates.
Because there are times when the matrix log function doesn't return
a valid log (often due to branch cut issues, e.g. taking the log of
a CNOT gate), this commit adds an assertion that the resulting dense
CPTP operation is the correct one (i.e. the one given to convert(...)).
This assertion is only tested when the gate dimension is <= 16, so as
not to call .todense() on very large gates.

This commit also adds some commented out code that gives a non-working
first attempt at an intelligent determination of the unitary postfactor
used by LindbladOps.  If this functionality were made to work it should
fix the log branch cut issues since the log will be taken of a near-identity
matrix.