Updates argument order in objective function calls.
Updates the argument ordering in objectivefns.py, chi2fns.py, and
likelihoodfns.py so that all of the objective-function calls have
roughly the same argument ordering (e.g. 'comm' before 'memLimit',
'regularization' before 'penalties').  This commit also updates
the function ordering and naming in chi2fns.py and likelihoodfns.py
so that these modules have a more parallel structure (though they're
not perfectly symmetrical).  The idea here is just to present
a more consistent interface to the user.

Calls to these functions from elsewhere in pyGSTi still need to
be updated, and should be assumed to be broken until then.