Massively refactors objectivefns.py, adding "raw" obj. functions.
In an effort to make the objective functions cleaner and easier to
maintain and extend, this commit separates the model-based part of
an objective function from the function-of-probabilities part, so
that now the typical factorization of an objective f as the
composition of g and probs functions:

f(paramvec) = g(probs(paramvec))

is more explicit.  The class hierarchy for objective functions
has been updated so that RawObjectiveFunction subclasses implement
the 'g' function and MDSObjectiveFunction subclasses contain
a 'probs'-implementing model as well as a raw objective function.
(MDS = "model and data set").  This division makes the
raw-objective-function code cleaner, and allows for better
code reuse.  Raw objective functions can have 'regularization'
parameters, e.g. minProbClip, and TimeIndependentMDSObjectiveFunction
object can have (model-based) penalties as well.

This commit also renames the methods of objective function classes
consistently and more specifically, creating 'lsvec' methods for the
least-squares objective function vector, etc.

This update, in addition to providing better internal organization,
allows the lightweight utility functions in chi2fns.py and
likelihoodfns.py (e.g. 'chi2fn') to be implemented by creating and
calling the methods of a raw objective function, thus consolidating
*all* the objective-function-implementing code into a single place.

General testing still needs to be done, and name changes throughout
the rest of pygsti still need to be make.  Renaming of variables/args
for PEP8 compliance still needs to happen.