Gets new objective function framework up and running.
Fixes lots of minor bugs/issues from basic testing using the recently
updated objective functions.  We compared the new objective functions
(the values, jacobian, and hessians) with the old functions (pre-
refactoring) and ensured that they gave the same results.  At this
point it seems that the objective functions are basically working,
and can be used to run GST, etc - EXCEPT the reports are still
broken (partly in need of objective-fn updates, but more so because
of the results object updates).  We'll wait until we merge in the
naming updates before tacking the reports.