Renames report figure factories, extras, and replib.
This commit updates fuction and argument naming to be PEP8
compliant in 3 main areas:
1) figure factory functions in report/section/*.py files.
   These functions still name 'workspace' as their first arg,
   so we ignore this naming warning in these files.
2) the extras/* files (we just didn't get to these earlier)
3) renamings performed in parallel in slowreplib.y and fastreplib.pyx.
   These modules still use captial-letter prefixes (e.g. "DM_")
   for functions, and we ignore the correspodning warning in
   slowreplib.py (fastreplib.pyx isn't linted currently)

Apart from the ignored warnings mentioned above, AND
our ignoring a couple naming warnings in legacyio.py
(so we can have classes like "dummy_GateString" and
 function names like "GateString_setstate"), this commit
fixes all of the linting warnings including the naming
warnings we want to comply with (all those except
"N806: variable in function should be lowercase").

Attempts were made not to break anything - especially
in the case of the report section updates which
required parallel re-naming in the HTML template files -
so everything should be close to working.  However,
nothing was tested at this point, so there's likely
some work needed to get everything back up and running.

Note that this commit also updates the .flake8 file to
reflect the naming errors we ignore globally and in
certain files as stated above.