Adds AdvancedOptions class and renames advanced options.
Adds advanced options classes to longsequence.py (maybe move later?)
which are essentially validated dicts so that users get immediate
feedback (a ValueError) when an unrecognized option is used.

This commit also renames the advanced options to use the
lowercase_with_underscores convention, just to give it uniformity
(previously we used a mix of all sorts of naming conventions).