Updates reporting framework to work with new objective fns & results.
Recent changes to the way objective functions are stored (namely,
they're held as self-contained objects now, built with "builder"
objects) and the ModelEstimateResults object (namely, the 'parameters'
of an estimate no longer hold copies of most of the protocol info)
broke the reporting framework.  This commit fixes that, with the
exception of getting all the caching to work again.  Places where
caching may be used but currently isn't is marked with CACHE
arguments, now set to None, which could be ComputationCaches passed
to objective functions.

This commit also update naming used in the reports, attempting to
replace "string" with "circuit" and "gs"/"gateset" with "mdl"/"model",
etc.  Reports seem to take longer than they should - but a standard
report can now be created successfully again.

Note: a future improvement to the reports would be to more
gracefully handle the case when there are no gauge optimizations
in the results object.  In this case, it would be nice to just leave
off the gauge-variant tabs/quantities of the report instead of having
empty/invisible tables with captions in weird places.